Add time zone name to date selection
This commit adds the time zone id to the label of the date chooser. This is the only way any confusion from users from different time zones might be avoided.
Vaadin 8 does not offer an option for localization an thus provides no possibility to retrieve the client time zone.

A more complex approach could include the parsing and localization of the client IP address but this is not considered necessary at the moment.